Assignments/Assignment 4/pyproject.toml
[project]
name = "assignment-4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dnspython>=2.8.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "pymongo[srv]>=4.15.2",
    "python-dotenv>=1.1.1",
]
